@@ -134,30 +134,28 @@
     "@types/flowjs": "^2.13.9",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "^2.0.10",
-    "@types/jquery": "^3.5.9",
+    "@types/jquery": "^3.5.16",
     "@types/js-beautify": "^1.13.3",
-    "@types/jstree": "^3.3.41",
-    "@types/leaflet": "^1.7.11",
+    "@types/leaflet": "^1.8.0",
     "@types/leaflet-polylinedecorator": "^1.6.1",
     "@types/leaflet-providers": "^1.2.1",
     "@types/leaflet.gridlayer.googlemutant": "^0.4.6",
     "@types/leaflet.markercluster": "^1.5.1",
-    "@types/lodash": "^4.14.177",
-    "@types/marked": "^4.0.3",
-    "@types/mousetrap": "^1.6.0",
-    "@types/node": "~15.14.9",
+    "@types/lodash": "^4.14.192",
+    "@types/marked": "^4.0.8",
+    "@types/node": "~18.15.11",
     "@types/raphael": "^2.3.2",
     "@types/react": "17.0.37",
     "@types/react-dom": "17.0.11",
     "@types/systemjs": "6.1.1",
     "@types/tinycolor2": "^1.4.3",
     "@types/tooltipster": "^0.0.31",
-    "@typescript-eslint/eslint-plugin": "5.43.0",
-    "@typescript-eslint/parser": "5.43.0",
+    "@typescript-eslint/eslint-plugin": "5.57.0",
+    "@typescript-eslint/parser": "5.57.0",
     "compression-webpack-plugin": "^10.0.0",
     "directory-tree": "^3.5.1",
     "glob": "^7.2.0",
-    "eslint": "^8.28.0",
+    "eslint": "^8.37.0",
     "eslint-plugin-import": "latest",
     "eslint-plugin-jsdoc": "latest",
     "eslint-plugin-prefer-arrow": "latest",
@@ -174,8 +172,8 @@
     "protractor": "~7.0.0",
     "raw-loader": "^4.0.2",
     "ts-node": "^10.9.1",
-    "typescript": "~4.8.4",
-    "webpack": "^5.75.0"
+    "typescript": "~4.9.5",
+    "webpack": "^5.77.0"
   },
   "resolutions": {
     "@types/react": "17.0.37",
