@@ -60,11 +60,11 @@ public class OtherPageElements extends AbstractBasePage {
     private static final String MARKS_CHECKBOX = "//mat-row[contains (@class,'mat-selected')]//mat-checkbox[contains(@class, 'checked')]";
     private static final String SELECT_ALL_CHECKBOX = "//thead//mat-checkbox";
     private static final String ALL_ENTITY = "//tbody/mat-row";
-    private static final String EDIT_PENCIL_BTN = "//mat-drawer-content//tb-entity-details-panel//mat-icon[contains(text(),'edit')]/ancestor::button";
+    private static final String EDIT_PENCIL_BTN = "//tb-entity-details-panel//mat-icon[contains(text(),'edit')]/ancestor::button";
     private static final String ENTITY_GROUP_EDIT_PENCIL_BTN = "//tb-entity-details-panel//mat-icon[contains(text(),'edit')]/ancestor::button";
     private static final String NAME_FIELD_EDIT_VIEW = "//input[@formcontrolname='name']";
     private static final String HEADER_NAME_VIEW = "//header//div[@class='tb-details-title']/span";
-    private static final String DONE_BTN_EDIT_VIEW = "//mat-drawer-content//mat-icon[contains(text(),'done')]/ancestor::button";
+    private static final String DONE_BTN_EDIT_VIEW = "//mat-icon[contains(text(),'done')]/ancestor::button";
     private static final String ENTITY_GROUP_DONE_BTN_EDIT_VIEW = "//mat-icon[contains(text(),'done')]/ancestor::button";
     private static final String DESCRIPTION_ENTITY_VIEW = "//mat-drawer-container[contains(@class,'has-open')]//textarea";
     private static final String DESCRIPTION_ADD_ENTITY_VIEW = "//mat-dialog-container//textarea";
