@@ -41,7 +41,7 @@ import org.testng.annotations.Test;
 import org.thingsboard.server.msa.ui.base.AbstractDriverBaseTest;
 import org.thingsboard.server.msa.ui.pages.CustomerPageHelper;
 import org.thingsboard.server.msa.ui.pages.LoginPageHelper;
-import org.thingsboard.server.msa.ui.pages.SideBarMenuViewElements;
+import org.thingsboard.server.msa.ui.pages.SideBarMenuViewHelper;
 
 import static org.thingsboard.server.msa.ui.base.AbstractBasePage.random;
 import static org.thingsboard.server.msa.ui.utils.Const.EMPTY_CUSTOMER_MESSAGE;
@@ -50,14 +50,14 @@ import static org.thingsboard.server.msa.ui.utils.Const.SAME_NAME_WARNING_CUSTOM
 
 public class CreateCustomerTest extends AbstractDriverBaseTest {
 
-    private SideBarMenuViewElements sideBarMenuView;
+    private SideBarMenuViewHelper sideBarMenuView;
     private CustomerPageHelper customerPage;
     private String customerName;
 
     @BeforeClass
     public void login() {
         new LoginPageHelper(driver).authorizationTenant();
-        sideBarMenuView = new SideBarMenuViewElements(driver);
+        sideBarMenuView = new SideBarMenuViewHelper(driver);
         customerPage = new CustomerPageHelper(driver);
     }
 
@@ -76,7 +76,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
     public void createCustomer() {
         String customerName = ENTITY_NAME + random();
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.plusBtn().click();
         customerPage.addCustomerViewEnterName(customerName);
         customerPage.addBtnC().click();
@@ -97,7 +97,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
         String email = "email@mail.com";
         String number = "12015550123";
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.plusBtn().click();
         customerPage.addCustomerViewEnterName(customerName);
         customerPage.selectCountryAddEntityView();
@@ -138,7 +138,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
     @Test(priority = 20, groups = "smoke")
     @Description("Add customer without the name")
     public void createCustomerWithoutName() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.plusBtn().click();
 
         Assert.assertFalse(customerPage.addBtnV().isEnabled());
@@ -149,7 +149,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
     @Test(priority = 20, groups = "smoke")
     @Description("Create customer only with spase in name")
     public void createCustomerWithOnlySpace() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.plusBtn().click();
         customerPage.addCustomerViewEnterName(Keys.SPACE);
         customerPage.addBtnC().click();
@@ -166,7 +166,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
     @Test(priority = 20, groups = "smoke")
     @Description("Create a customer with the same name")
     public void createCustomerSameName() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.setCustomerName();
         String customerName = customerPage.getCustomerName();
         customerPage.plusBtn().click();
@@ -187,7 +187,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
     public void createCustomerWithoutRefresh() {
         String customerName = ENTITY_NAME + random();
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.plusBtn().click();
         customerPage.addCustomerViewEnterName(customerName);
         customerPage.addBtnC().click();
@@ -204,7 +204,7 @@ public class CreateCustomerTest extends AbstractDriverBaseTest {
     public void documentation() {
         String urlPath = "/docs/pe/user-guide/ui/customers/";
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.setCustomerName();
         customerPage.customer(customerPage.getCustomerName()).click();
         customerPage.goToHelpPage();
