@@ -41,7 +41,7 @@ import org.testng.annotations.Test;
 import org.thingsboard.server.msa.ui.base.AbstractDriverBaseTest;
 import org.thingsboard.server.msa.ui.pages.CustomerPageHelper;
 import org.thingsboard.server.msa.ui.pages.LoginPageHelper;
-import org.thingsboard.server.msa.ui.pages.SideBarMenuViewElements;
+import org.thingsboard.server.msa.ui.pages.SideBarMenuViewHelper;
 
 import static org.thingsboard.server.msa.ui.base.AbstractBasePage.random;
 import static org.thingsboard.server.msa.ui.utils.Const.ENTITY_NAME;
@@ -49,7 +49,7 @@ import static org.thingsboard.server.msa.ui.utils.Const.OWNER_NOT_SELECTED_ERROR
 import static org.thingsboard.server.msa.ui.utils.EntityPrototypes.defaultCustomerPrototype;
 
 public class CustomerChangeOwnerTest extends AbstractDriverBaseTest {
-    private SideBarMenuViewElements sideBarMenuView;
+    private SideBarMenuViewHelper sideBarMenuView;
     private CustomerPageHelper customerPage;
     private final String title = ENTITY_NAME + random();
     private final String title1 = ENTITY_NAME + random() + '1';
@@ -57,7 +57,7 @@ public class CustomerChangeOwnerTest extends AbstractDriverBaseTest {
     @BeforeClass
     public void login() {
         new LoginPageHelper(driver).authorizationTenant();
-        sideBarMenuView = new SideBarMenuViewElements(driver);
+        sideBarMenuView = new SideBarMenuViewHelper(driver);
         customerPage = new CustomerPageHelper(driver);
     }
 
@@ -74,13 +74,11 @@ public class CustomerChangeOwnerTest extends AbstractDriverBaseTest {
         testRestClient.postCustomer(defaultCustomerPrototype(title));
         testRestClient.postCustomer(defaultCustomerPrototype(title1));
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.checkBox(title1).click();
         customerPage.changeOwnerBtn().click();
         customerPage.changeOwner(title);
-        customerPage.waitUntilCustomerNotVisible(title1);
         customerPage.manageCustomerGroupsBtn(title).click();
-        customerPage.entity("All").click();
 
         Assert.assertNotNull(customerPage.entity(title1));
         Assert.assertTrue(customerPage.entity(title1).isDisplayed());
@@ -93,7 +91,7 @@ public class CustomerChangeOwnerTest extends AbstractDriverBaseTest {
     public void changeOwnerWithoutName() {
         testRestClient.postCustomer(defaultCustomerPrototype(title));
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.checkBox(title).click();
         customerPage.changeOwnerBtn().click();
         customerPage.changeOwnerViewField().click();
@@ -113,7 +111,7 @@ public class CustomerChangeOwnerTest extends AbstractDriverBaseTest {
     public void changeOwnerWithOnlySpace() {
         testRestClient.postCustomer(defaultCustomerPrototype(title));
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.checkBox(title).click();
         customerPage.changeOwnerBtn().click();
         customerPage.changeOwnerViewField().sendKeys(" ");
@@ -131,14 +129,12 @@ public class CustomerChangeOwnerTest extends AbstractDriverBaseTest {
         testRestClient.postCustomer(defaultCustomerPrototype(title1));
         testRestClient.postCustomer(defaultCustomerPrototype(title3));
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.checkBox(title1).click();
         customerPage.checkBox(title3).click();
         customerPage.changeOwnerBtn().click();
         customerPage.changeOwner(title);
-        customerPage.waitUntilCustomerNotVisible(title1);
         customerPage.manageCustomerGroupsBtn(title).click();
-        customerPage.entity("All").click();
 
         Assert.assertNotNull(customerPage.entity(title1));
         Assert.assertNotNull(customerPage.entity(title3));
