@@ -39,7 +39,7 @@ import org.testng.annotations.Test;
 import org.thingsboard.server.msa.ui.base.AbstractDriverBaseTest;
 import org.thingsboard.server.msa.ui.pages.CustomerPageHelper;
 import org.thingsboard.server.msa.ui.pages.LoginPageHelper;
-import org.thingsboard.server.msa.ui.pages.SideBarMenuViewElements;
+import org.thingsboard.server.msa.ui.pages.SideBarMenuViewHelper;
 
 import static org.thingsboard.server.msa.ui.base.AbstractBasePage.random;
 import static org.thingsboard.server.msa.ui.utils.Const.ENTITY_NAME;
@@ -47,13 +47,13 @@ import static org.thingsboard.server.msa.ui.utils.EntityPrototypes.defaultCustom
 
 public class DeleteSeveralCustomerTest extends AbstractDriverBaseTest {
 
-    private SideBarMenuViewElements sideBarMenuView;
+    private SideBarMenuViewHelper sideBarMenuView;
     private CustomerPageHelper customerPage;
 
     @BeforeClass
     public void login() {
         new LoginPageHelper(driver).authorizationTenant();
-        sideBarMenuView = new SideBarMenuViewElements(driver);
+        sideBarMenuView = new SideBarMenuViewHelper(driver);
         customerPage = new CustomerPageHelper(driver);
     }
 
@@ -67,7 +67,7 @@ public class DeleteSeveralCustomerTest extends AbstractDriverBaseTest {
         testRestClient.postCustomer(defaultCustomerPrototype(title1));
         testRestClient.postCustomer(defaultCustomerPrototype(title2));
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.deleteSelected(2);
         customerPage.refreshBtn().click();
 
@@ -81,7 +81,7 @@ public class DeleteSeveralCustomerTest extends AbstractDriverBaseTest {
     @Description("Remove several customers by mark all the Customers on the page by clicking in the topmost checkbox " +
             "and then clicking on the trash icon in the menu that appears")
     public void selectAllCustomers() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.selectAllCheckBox().click();
         jsClick(customerPage.deleteSelectedBtn());
 
@@ -101,7 +101,7 @@ public class DeleteSeveralCustomerTest extends AbstractDriverBaseTest {
         testRestClient.postCustomer(defaultCustomerPrototype(title1));
         testRestClient.postCustomer(defaultCustomerPrototype(title2));
 
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.deleteSelected(2);
 
         Assert.assertTrue(customerPage.customerIsNotPresent(title1));
