@@ -39,17 +39,17 @@ import org.testng.annotations.Test;
 import org.thingsboard.server.msa.ui.base.AbstractDriverBaseTest;
 import org.thingsboard.server.msa.ui.pages.CustomerPageHelper;
 import org.thingsboard.server.msa.ui.pages.LoginPageHelper;
-import org.thingsboard.server.msa.ui.pages.SideBarMenuViewElements;
+import org.thingsboard.server.msa.ui.pages.SideBarMenuViewHelper;
 
 public class ManageCustomersDevicesTest extends AbstractDriverBaseTest {
-    private SideBarMenuViewElements sideBarMenuView;
+    private SideBarMenuViewHelper sideBarMenuView;
     private CustomerPageHelper customerPage;
-    private final String iconText = ": Device groups";
+    private final String iconText = ": Devices";
 
     @BeforeClass
     public void login() {
         new LoginPageHelper(driver).authorizationTenant();
-        sideBarMenuView = new SideBarMenuViewElements(driver);
+        sideBarMenuView = new SideBarMenuViewHelper(driver);
         customerPage = new CustomerPageHelper(driver);
     }
 
@@ -58,11 +58,11 @@ public class ManageCustomersDevicesTest extends AbstractDriverBaseTest {
     @Test(groups = "smoke")
     @Description("Open manage window by right corner btn")
     public void openWindowByRightCornerBtn() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.setCustomerName();
         customerPage.manageCustomersDeviceGroupsBtn(customerPage.getCustomerName()).click();
 
-        Assert.assertTrue(urlContains("deviceGroups"));
+        Assert.assertTrue(urlContains("devices"));
         Assert.assertNotNull(customerPage.customerUserIconHeader());
         Assert.assertTrue(customerPage.customerUserIconHeader().isDisplayed());
         Assert.assertTrue(customerPage.customerUserIconHeader().getText().contains(customerPage.getCustomerName() + iconText));
@@ -74,12 +74,12 @@ public class ManageCustomersDevicesTest extends AbstractDriverBaseTest {
     @Test(groups = "smoke")
     @Description("Open manage window by btn in entity view")
     public void openWindowByView() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.setCustomerName();
         customerPage.entity(customerPage.getCustomerName()).click();
         jsClick(customerPage.manageCustomersDeviceGroupsBtnView());
 
-        Assert.assertTrue(urlContains("deviceGroups"));
+        Assert.assertTrue(urlContains("devices"));
         Assert.assertNotNull(customerPage.customerUserIconHeader());
         Assert.assertTrue(customerPage.customerUserIconHeader().isDisplayed());
         Assert.assertTrue(customerPage.customerUserIconHeader().getText().contains(customerPage.getCustomerName() + iconText));
