@@ -39,18 +39,18 @@ import org.testng.annotations.Test;
 import org.thingsboard.server.msa.ui.base.AbstractDriverBaseTest;
 import org.thingsboard.server.msa.ui.pages.CustomerPageHelper;
 import org.thingsboard.server.msa.ui.pages.LoginPageHelper;
-import org.thingsboard.server.msa.ui.pages.SideBarMenuViewElements;
+import org.thingsboard.server.msa.ui.pages.SideBarMenuViewHelper;
 
 public class ManageCustomersUsersTest extends AbstractDriverBaseTest {
 
-    private SideBarMenuViewElements sideBarMenuView;
+    private SideBarMenuViewHelper sideBarMenuView;
     private CustomerPageHelper customerPage;
-    private final String iconText = ": User groups";
+    private final String iconText = ": Users";
 
     @BeforeClass
     public void login() {
         new LoginPageHelper(driver).authorizationTenant();
-        sideBarMenuView = new SideBarMenuViewElements(driver);
+        sideBarMenuView = new SideBarMenuViewHelper(driver);
         customerPage = new CustomerPageHelper(driver);
     }
 
@@ -59,11 +59,11 @@ public class ManageCustomersUsersTest extends AbstractDriverBaseTest {
     @Test(groups = "smoke")
     @Description("Open manage window by right corner btn")
     public void openWindowByRightCornerBtn() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.setCustomerName();
         jsClick(customerPage.manageCustomersUserBtn(customerPage.getCustomerName()));
 
-        Assert.assertTrue(urlContains("userGroups"));
+        Assert.assertTrue(urlContains("users"));
         Assert.assertNotNull(customerPage.customerUserIconHeader());
         Assert.assertTrue(customerPage.customerUserIconHeader().isDisplayed());
         Assert.assertTrue(customerPage.customerUserIconHeader().getText().contains(customerPage.getCustomerName() + iconText));
@@ -75,12 +75,12 @@ public class ManageCustomersUsersTest extends AbstractDriverBaseTest {
     @Test(groups = "smoke")
     @Description("Open manage window by btn in entity view")
     public void openWindowByView() {
-        sideBarMenuView.goToAllCustomerGroupBtn();
+        sideBarMenuView.goToAllCustomers();
         customerPage.setCustomerName();
         customerPage.entity(customerPage.getCustomerName()).click();
         jsClick(customerPage.manageCustomersUserGroupsBtnView());
 
-        Assert.assertTrue(urlContains("userGroups"));
+        Assert.assertTrue(urlContains("users"));
         Assert.assertNotNull(customerPage.customerUserIconHeader());
         Assert.assertTrue(customerPage.customerUserIconHeader().isDisplayed());
         Assert.assertTrue(customerPage.customerUserIconHeader().getText().contains(customerPage.getCustomerName() + iconText));
