@@ -47,6 +47,9 @@ import { EntityInfoData } from '@shared/models/entity.models';
 import { UserPermissionsService } from '@core/http/user-permissions.service';
 import { Operation } from '@shared/models/security.models';
 import { EntityId } from '@app/shared/models/id/entity-id';
+import { HomeDialogsService } from '@home/dialogs/home-dialogs.service';
+import { CreateEntityGroupFunction } from '@shared/components/group/entity-group-list.component';
+import { map } from 'rxjs/operators';
 
 export interface OwnerAndGroupsData {
   owner?: EntityId | EntityInfoData;
@@ -90,10 +93,13 @@ export class OwnerAndGroupsComponent extends PageComponent implements OnInit, Co
   private ownerDisabled = false;
   private groupsDisabled = false;
 
+  createGroupFunction: CreateEntityGroupFunction;
+
   constructor(protected store: Store<AppState>,
               private translate: TranslateService,
               private fb: UntypedFormBuilder,
-              private userPermissionsService: UserPermissionsService) {
+              private userPermissionsService: UserPermissionsService,
+              private homeDialogs: HomeDialogsService) {
     super(store);
   }
 
@@ -102,6 +108,21 @@ export class OwnerAndGroupsComponent extends PageComponent implements OnInit, Co
       this.ownerDisabled = !this.userPermissionsService.hasGenericPermissionByEntityGroupType(Operation.CHANGE_OWNER, this.entityType);
       this.groupsDisabled = !this.userPermissionsService.hasGenericEntityGroupTypePermission(Operation.ADD_TO_GROUP, this.entityType) ||
                             !this.userPermissionsService.hasGenericEntityGroupTypePermission(Operation.REMOVE_FROM_GROUP, this.entityType);
+    }
+    if (this.userPermissionsService.hasGenericEntityGroupTypePermission(Operation.CREATE, this.entityType)) {
+      this.createGroupFunction = (groupType, groupName, ownerId) =>
+        this.homeDialogs.createEntityGroup(groupType, groupName, ownerId).pipe(
+        map((result) => {
+          if (result && result.entityGroup) {
+            return {
+              id: result.entityGroup.id,
+              name: result.entityGroup.name
+            };
+          } else {
+            return null;
+          }
+        }
+      ));
     }
 
     this.ownerAndGroupsFormGroup = this.fb.group({
