@@ -37,8 +37,7 @@ import { AddEntitiesToEdgeDialogComponent } from '@home/dialogs/add-entities-to-
 import { SelectOwnerDialogComponent } from '@home/dialogs/select-owner-dialog.component';
 import { SelectEntityGroupDialogComponent } from '@home/dialogs/select-entity-group-dialog.component';
 import { ShareEntityGroupDialogComponent } from '@home/dialogs/share-entity-group-dialog.component';
-import { AddEntityGroupsToEdgeDialogComponent } from "@home/dialogs/add-entity-groups-to-edge-dialog.component";
-import { ManageOwnerAndGroupsDialogComponent } from '@home/dialogs/manage-owner-and-groups-dialog.component';
+import { AddEntityGroupsToEdgeDialogComponent } from '@home/dialogs/add-entity-groups-to-edge-dialog.component';
 
 @NgModule({
   declarations:
@@ -47,8 +46,7 @@ import { ManageOwnerAndGroupsDialogComponent } from '@home/dialogs/manage-owner-
     SelectEntityGroupDialogComponent,
     ShareEntityGroupDialogComponent,
     AddEntityGroupsToEdgeDialogComponent,
-    AddEntitiesToEdgeDialogComponent,
-    ManageOwnerAndGroupsDialogComponent
+    AddEntitiesToEdgeDialogComponent
   ],
   imports: [
     CommonModule,
@@ -59,8 +57,7 @@ import { ManageOwnerAndGroupsDialogComponent } from '@home/dialogs/manage-owner-
     SelectEntityGroupDialogComponent,
     ShareEntityGroupDialogComponent,
     AddEntityGroupsToEdgeDialogComponent,
-    AddEntitiesToEdgeDialogComponent,
-    ManageOwnerAndGroupsDialogComponent
+    AddEntitiesToEdgeDialogComponent
   ],
   providers: [
     HomeDialogsService
