@@ -56,7 +56,7 @@ import { EntityGroupInfo } from '@shared/models/entity-group.models';
 import { BaseData, HasId } from '@shared/models/base-data';
 import { UserPermissionsService } from '@core/http/user-permissions.service';
 import { BreakpointObserver, BreakpointState } from '@angular/cdk/layout';
-import { OwnerAndGroupsData } from '@shared/components/group/owner-and-groups.component';
+import { OwnerAndGroupsData } from '@home/components/group/owner-and-groups.component';
 import { MediaBreakpoints } from '@shared/models/constants';
 import { StepperSelectionEvent } from '@angular/cdk/stepper';
 import { getCurrentAuthUser } from '@core/auth/auth.selectors';
@@ -223,15 +223,8 @@ export class AddUserDialogComponent extends DialogComponent<AddUserDialogCompone
         this.user.authority = Authority.CUSTOMER_USER;
         this.user.customerId = targetOwnerId as CustomerId;
       }
-
-      let saveUserObservable: Observable<User>;
-      if (this.isSysAdmin) {
-        saveUserObservable = this.userService.saveUser(this.user, sendActivationEmail);
-      } else {
-        const entityGroupIds = targetOwnerAndGroups.groups.map(group => group.id.id);
-        saveUserObservable = this.userService.saveUser(this.user, sendActivationEmail, entityGroupIds);
-      }
-      saveUserObservable.subscribe(
+      const entityGroupIds = targetOwnerAndGroups.groups.map(group => group.id.id);
+      this.userService.saveUser(this.user, sendActivationEmail, entityGroupIds).subscribe(
         (user) => {
           if (this.activationMethod === ActivationMethod.DISPLAY_ACTIVATION_LINK) {
             this.userService.getActivationLink(user.id.id).subscribe(
